Adjust config tuist
