+++ .mise.toml
@@ -1,4 +1,4 @@
 [tools]
 tuist = "4.26.0"
-swiftlint = "0.56.2"
-swiftformat = "0.54.3"
+swiftlint = "0.57.0"
+swiftformat = "0.54.4"
