+++ Tuist/Config.swift
@@ -1,0 +1,3 @@
+import ProjectDescription
+
+let config = Config()
